requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0,<2.0.0
Pillow>=8.3.0
matplotlib>=3.4.0
tensorboard>=2.6.0
tqdm>=4.62.0
opencv-python>=4.5.0
scikit-image>=0.18.0
albumentations>=1.0.0
wandb>=0.12.0
